fix sign_up issue
