+++ Web Application Codes/web/src/components/Ride.js
@@ -1,10 +1,14 @@
 import React from 'react';
+import Nav from './Nav';
+import Footer from './Footer';
 import MapsPage from './MapsPage';
 
 function Ride() {
 	return (
-		<div className='bodDiv'>
-			{/* <div className='formDiv'>
+		<div>
+            <Nav />
+			<div className='bodDiv'>
+				{/* <div className='formDiv'>
                 <h1>Select Ride Type</h1>
                 <form>
                     <input type='text' placeholder='enter you pick up location' />
@@ -12,14 +16,16 @@ function Ride() {
                     <input type='submit' value='order now' className='submit'/>
                 </form>
             </div> */}
-			{/* <div className='maps'> */}
-			{/* <h1>Select Ride Type</h1>
+				{/* <div className='maps'> */}
+				{/* <h1>Select Ride Type</h1>
                 <form>
                     <input type='text' placeholder='enter you pick up location' />
                     <input type='text' placeholder='enter you destination'/>
                 </form> */}
-			{/* </div> */}
-			{/* <MapsPage /> */}
+				{/* </div> */}
+				{/* <MapsPage /> */}
+			</div>
+            <Footer />
 		</div>
 	);
 }
