+++ Web Application Codes/web/src/css/sign_up_new.css
@@ -1,0 +1,79 @@
+#signup-top-nav {
+	background-color: #00000043;
+	height: 10vh;
+	width: 100%;
+	display: flex;
+	flex-direction: row;
+	flex-wrap: nowrap;
+	justify-content: space-around;
+	align-items: center;
+}
+
+#signup-top-nav div {
+	flex: 0 0 40%;
+}
+
+#signup-top-list {
+	flex: 0 0 30%;
+	color: white;
+	display: flex;
+	flex-direction: row;
+	flex-wrap: nowrap;
+	justify-content: space-between;
+	height: 100%;
+}
+
+#signup-top-list li {
+	display: block;
+	list-style-type: none;
+	height: 100%;
+	flex: 0 0 30%;
+	text-align: center;
+	padding-top: 6%;
+}
+
+#signup-top-list li::after {
+	content: '';
+	display: block;
+	color: #000000;
+	width: 0;
+	height: 2px;
+	border: 0px solid #fff;
+	transition: all 0.2s;
+}
+
+#signup-top-list li:hover::after {
+	border-top: 4px solid #fff;
+	width: 100%;
+}
+
+a {
+	color: white;
+	font-size: 120%;
+	font-weight: 400;
+	display: block;
+	width: 100%;
+	height: 100%;
+}
+
+.signup-page-background {
+	background: url(./action-adult-blur-car-620335_i.png);
+	background-size: cover;
+	background-repeat: no-repeat;
+	height: 100vh;
+	width: 100vw;
+	overflow: hidden;
+}
+.signup-page-foreground {
+	height: inherit;
+	background-image: linear-gradient(180deg, #0000003d, #000000);
+}
+
+#signup-top-list li {
+	margin-bottom: 10px;
+	color: #fff;
+}
+
+.p span {
+	color: #ea355c;
+}
