+++ Web Application Codes/web/src/css/welcome.css
@@ -1,0 +1,102 @@
+.topNav {
+	width: 100%;
+	height: 10vh;
+	align-items: center;
+	background: rgba(0, 0, 0, 0.4);
+	display: inline-flex;
+}
+
+.topLogo svg {
+	background-color: azure;
+	padding: 1%;
+	border: rgb(241, 54, 132) 2px solid;
+	box-shadow: 3px 4px 6px rgba(0, 0, 0, 0.3);
+	border-radius: 10px;
+	width: 10%;
+	line-height: 50px;
+	height: 7vh;
+	display: inline;
+}
+
+.topHeader {
+	line-height: 50px;
+	color: azure;
+	margin: 0 2%;
+}
+
+.topLogo {
+	display: inline-flex;
+	margin-left: 7%;
+	width: 50%;
+}
+
+
+.topList {
+	height: 100%;
+	display: flex;
+	text-align: center;
+	list-style-type: none;
+	flex-wrap: nowrap;
+	margin-left: 10%;
+}
+.topList li {
+	display: inline-flex;
+	list-style-type: none;
+	margin-right: 25%;
+	color: azure;
+	height: inherit;
+}
+.topList li a {
+	height: 100%;
+	text-decoration: none;
+	color: azure;
+	line-height: 10vh;
+}
+
+a img {
+	color: azure;
+	width: 20%;
+}
+
+.welcome {
+	text-align: center;
+	align-items: center;
+	width: 28%;
+	background-color: #fff;
+	margin-left: 37%;
+	margin-top: 7%;
+	padding: 3% 1%;
+	padding-top: 1.5%;
+	border-radius: 15px;
+}
+.welcome .topLogo {
+	padding: 5%;
+	text-align: center;
+	align-items: center;
+	width: 120%;
+	margin-left: 40%;
+}
+.welcomeWord {
+	margin: 2%;
+	color: rgb(241, 54, 132);
+	margin-bottom: 20%;
+}
+.welcome a {
+	background: rgb(241, 54, 132);
+	color: #fff;
+	padding: 3.5% 22%;
+	border-radius: 8px;
+}
+
+.signup-page-background {
+	background: url(./action-adult-blur-car-620335_i.png);
+	background-size: cover;
+	background-repeat: no-repeat;
+	height: 100vh;
+	width: 100vw;
+	overflow: hidden;
+}
+.signup-page-foreground {
+	height: inherit;
+	background-image: linear-gradient(180deg, #0000003d, #000000);
+}
